Extract engine export wrapper
export_video is a thin compatibility wrapper around convert that marks the operation as export. This moves it into engine_export.py while re-exporting export_video from mcp_video.engine and preserving the mcp_video.engine.convert patch point.

Constraint: Preserve export behavior, progress forwarding, two-pass args, and existing monkeypatch expectations around mcp_video.engine.convert.

Rejected: Move convert with export_video | convert remains the larger high-risk encoding surface and should be isolated separately.

Confidence: high

Scope-risk: narrow

Directive: Keep wrapper extractions separate from core encoding algorithm changes.

Tested: ruff check mcp_video/engine.py mcp_video/engine_export.py

Tested: ruff format --check mcp_video/engine.py mcp_video/engine_export.py

Tested: /opt/homebrew/bin/python3 export_video re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine_advanced.py::TestExportVideo tests/test_engine_advanced.py::TestTwoPassEncoding::test_export_video_two_pass tests/test_cli.py::TestCLIExport tests/test_server.py::TestVideoExportTool -q --tb=short

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short
